move utils in server
