+++ src/gitingest/config.py
@@ -7,4 +7,5 @@ MAX_DIRECTORY_DEPTH = 20  # Maximum depth of directory traversal
 MAX_FILES = 10_000  # Maximum number of files to process
 MAX_TOTAL_SIZE_BYTES = 500 * 1024 * 1024  # 500 MB
 
+OUTPUT_FILE_PATH = "digest.txt"
 TMP_BASE_PATH = Path("/tmp/gitingest")
