+++ src/server/main.py
@@ -10,9 +10,10 @@ from fastapi.staticfiles import StaticFiles
 from slowapi.errors import RateLimitExceeded
 from starlette.middleware.trustedhost import TrustedHostMiddleware
 
+from gitingest.config import TMP_BASE_PATH
 from server.routers import download, dynamic, index
-from server.server_config import DELETE_REPO_AFTER, TMP_BASE_PATH
-from server.server_utils import limiter
+from server.server_config import DELETE_REPO_AFTER, templates
+from server.server_utils import lifespan, limiter, rate_limit_exception_handler
 
 # Load environment variables from .env file
 load_dotenv()
@@ -42,9 +43,6 @@ else:
 
 # Add middleware to enforce allowed hosts
 app.add_middleware(TrustedHostMiddleware, allowed_hosts=allowed_hosts)
-
-# Set up template rendering
-templates = Jinja2Templates(directory="server/templates")
 
 
 @app.get("/health")
